modify to use typescript and grafana-sdk-mocks
