+++ .gitignore
@@ -1,36 +1,13 @@
 node_modules
-npm-debug.log
-coverage/
-.aws-config.json
-awsconfig
-/emails/dist
-/public_gen
-/tmp
-vendor/phantomjs/phantomjs
-
-docs/AWS_S3_BUCKET
-docs/GIT_BRANCH
-docs/VERSION
-docs/GITCOMMIT
-docs/changed-files
-docs/changed-files
-
-# locally required config files
-public/css/*.min.css
 
 # Editor junk
-*.sublime-workspace
 *.swp
 .idea/
 *.iml
 
-/data/*
-/bin/*
-
-conf/custom.ini
-fig.yml
-profile.cov
-grafana
 .notouch
 .DS_Store
 .tscache
+
+# don't commit every version
+dist/
