+++ src/plugin.json
@@ -12,11 +12,11 @@
     "logos": { },
     "links": [ ],
     "screenshots": [ ],
-    "version": "0.0.1",
-    "updated": "2017-04-05"
+    "version": "%VERSION%",
+    "updated": "%TODAY%"
   },
   "dependencies": {
-    "grafanaVersion": "4.6.x",
+    "grafanaVersion": "5.x",
     "plugins": [ ]
   }
 }
