+++ tsconfig.json
@@ -1,0 +1,23 @@
+{
+    "compilerOptions": {
+      "moduleResolution": "node",
+      "outDir": "./dist/",
+      "target": "es5",
+      "lib": ["es6", "dom"],
+      "rootDir": "src/",
+      "jsx": "react",
+      "module": "esnext",
+      "declaration": false,
+      "allowSyntheticDefaultImports": true,
+      "inlineSourceMap": false,
+      "sourceMap": true,
+      "noEmitOnError": false,
+      "emitDecoratorMetadata": false,
+      "experimentalDecorators": true,
+      "noImplicitReturns": true,
+      "noImplicitThis": false,
+      "noImplicitUseStrict":false,
+      "noImplicitAny": false,
+      "noUnusedLocals": true,
+    }
+}
